cpd_app/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
numpy
opencv-python-headless
Pillow
